+++ Code/Nodes/BackgroundManager.swift
